further updates for testing with pytest
